Change iterkeys() to keys() method for Python 2 and 3 compatibility.
Note that this is sub-optimal in that it triggers creation of a list of the
dict keys in Python 2 where an iterator would do. However, it is the only use
of iterkeys() in SOGcommand so it doesn't seem worth adding complexity or a
dependency on the six library to preserve the iterator use in Python 2.
Use Python 3 for speed and memory efficiency!